requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
coremltools==5.1.0
cycler==0.11.0
Cython==0.29.25
docker-pycreds==0.4.0
docstring-parser==0.13
fire==0.4.0
fonttools==4.28.1
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
imageio==2.11.0
importlib-metadata==4.8.2
joblib==1.1.0
jsonargparse==4.0.0
kaleido==0.2.1
kiwisolver==1.3.2
lightning-flash==0.6.0
llvmlite==0.37.0
loguru==0.5.3
Markdown==3.3.6
mpmath==1.2.1
multidict==5.2.0
networkx==2.6.3
numba==0.54.1
numpy==1.20.1
oauthlib==3.1.1
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==21.3
pandas==1.2.1
pathtools==0.1.2
Pillow
plotly==5.2.2
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyDeprecate==0.3.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytorch-lightning==1.5.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.1
scipy==1.7.2
sentry-sdk==1.5.0
setuptools-scm==6.3.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
sympy==1.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tomli==1.2.2
torch==1.8.0
torchaudio==0.8.0
torchmetrics==0.6.0
torchsampler @ https://github.com/ufoym/imbalanced-dataset-sampler/archive/master.zip
torchvision==0.9.0
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.0
urllib3==1.26.7
wandb==0.12.6
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
